// src/app/shared/model/BitacoraDTO.ts
export class BitacoraDTO {
    idIncidencia: number;
    catTipoIncidencia: number;
    idSistema: string;
    idAmbiente: string;
    idPantalla: number;
    usuario: string;
    ubicacion: string;
    accion: string;
    valorAnterior: string;
    valorActual: string;
    descripcionIncidencia: string;
    fechaIncidencia: Date;
}